Send mail message: fix mails not sent
Fix an issue where too many calls to mailBuilder.send were made, causing
an exception caused by the smtp server.
Fixes RM28522
