+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Improvements
 ## Bug Fixes
 - CRM/SALES : fix for contacts with mainPartner isProspect do not display
+- Mail message: fix an issue where some emails are not being sent in a batch.
 
 ## [5.3.6] - 2020-05-26
 ## Improvements
